sesssionContext 앱/웹 구분 없이 통일
